requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
boto3==1.34.137
botocore==1.34.137
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
dnspython==2.6.1
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
orjson==3.10.5
packaging==23.2
pandas==2.2.2
parso==0.8.3
pillow==10.4.0
pip==24.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
scipy==1.14.0
seaborn==0.13.2
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.48
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.12
websockets==12.0
wheel==0.43.0
